Detach python worker resolution into resolve() and a bringup crate
Worker launch config was resolved by Config::prepare(), welding host
probing onto the declarative config type, and each of the three bins
repeated the same prepare-then-wire dance.

Split the two concerns:
- Resolution is now a free `resolve(Config) -> SpecFactory` whose probe
  runs once under spawn_blocking; Config is inert declarative data.
- `SpecFactory` (was `PreparedConfig`) is named for its job: `build(addr)`
  mints a `Spec` once the late-bound bridge address is known. `Spec`
  fields are private, constructible only via the factory.
- New `worker-python-bringup` crate owns the compose; each bin collapses
  to a single `start(...)` call. `worker-remote-bringup` stays generic
  and python-free.

`PrepareError` -> `ResolveError`; `config/prepared.rs` moves to top-level
`resolve.rs` (resolution is no longer an extension of config). AGENTS.md
rule examples updated to the new names.
