+++ AGENTS.md
@@ -168,6 +168,6 @@ This section is used for the scratch updates, driven by our Agents.
 <rule>Use distinct error variants for distinct failure states instead of reusing a nearby transport error. Good: `return Err(SendActionError::WorkerProtocolClosed);` Bad: `return Err(SendActionError::ChannelClosed);` when protocol state was already closed before enqueue.</rule>
 <rule>Model closed or unavailable shared state with `Option<T>` under the lock instead of parallel boolean flags. Good: `Mutex<Option<SharedState>>` with `guard.take()` on shutdown. Bad: `SharedState { closed: bool, ... }` plus manual checks.</rule>
 <rule>When shared async state must be closed during unwind or normal teardown, prefer a synchronous mutex plus a drop guard if the lock is never held across `.await`. Good: `let _guard = SharedStateDropGuard::new(Arc::clone(&shared));` with `Arc<std::sync::Mutex<Option<SharedState>>>`. Bad: async-only teardown paths that skip cleanup on panic.</rule>
-<rule>Don't run blocking filesystem/IO on the async runtime in a hot path; resolve and cache the materialized values once up front (e.g. an async `prepare()` that runs the probing under `spawn_blocking`) and have the hot path read the cache. Good: `Config::prepare().await` yields a `PreparedConfig` that a pure `prepare_spawn_params` reads. Bad: re-running `is_dir()`/`current_dir()`/`var("PATH")` on every spawn inside the trait impl.</rule>
-<rule>Prefer fallible paths in failure-tolerance scopes, resolving them before the values are passed into hot or infallible paths; don't `unwrap`/`expect`/panic inside an infallible trait method or hot path. Good: `std::env::current_dir().map_err(PrepareError::CurrentDir)?` inside an async `prepare()`. Bad: `std::env::current_dir().expect("...")` inside `prepare_spawn_params`.</rule>
+<rule>Don't run blocking filesystem/IO on the async runtime in a hot path; resolve and cache the materialized values once up front (e.g. an async preparation step that runs the probing under `spawn_blocking`) and have the hot path read the cache. Good: `waymark_worker_python::resolve(config).await` yields a `SpecFactory` that a pure `prepare_spawn_params` reads. Bad: re-running `is_dir()`/`current_dir()`/`var("PATH")` on every spawn inside the trait impl.</rule>
+<rule>Prefer fallible paths in failure-tolerance scopes, resolving them before the values are passed into hot or infallible paths; don't `unwrap`/`expect`/panic inside an infallible trait method or hot path. Good: `std::env::current_dir().map_err(ResolveError::CurrentDir)?` inside an async `resolve()`. Bad: `std::env::current_dir().expect("...")` inside `prepare_spawn_params`.</rule>
 </code_feedback>
