+++ Cargo.toml
@@ -108,6 +108,7 @@ waymark-worker-process = { path = "crates/lib/worker-process" }
 waymark-worker-process-pool = { path = "crates/lib/worker-process-pool" }
 waymark-worker-process-spec = { path = "crates/lib/worker-process-spec" }
 waymark-worker-python = { path = "crates/lib/worker-python" }
+waymark-worker-python-bringup = { path = "crates/lib/worker-python-bringup" }
 waymark-worker-remote-bridge-bringup = { path = "crates/lib/worker-remote-bridge-bringup" }
 waymark-worker-remote-bridge-service = { path = "crates/lib/worker-remote-bridge-service" }
 waymark-worker-remote-bringup = { path = "crates/lib/worker-remote-bringup" }
