+++ crates/bin/smoke/Cargo.toml
@@ -19,7 +19,7 @@ waymark-runner-state = { workspace = true }
 waymark-smoke-sources = { workspace = true }
 waymark-worker-process-spec = { workspace = true }
 waymark-worker-python = { workspace = true }
-waymark-worker-remote-bringup = { workspace = true }
+waymark-worker-python-bringup = { workspace = true }
 waymark-worker-remote-pool = { workspace = true }
 waymark-workflow-registry-backend = { workspace = true }
