+++ crates/bin/start-workers/Cargo.toml
@@ -21,7 +21,7 @@ waymark-scheduler-loop-core = { workspace = true }
 waymark-tokio-metrics-bringup = { workspace = true }
 waymark-webapp-bringup = { workspace = true }
 waymark-worker-python = { workspace = true }
-waymark-worker-remote-bringup = { workspace = true }
+waymark-worker-python-bringup = { workspace = true }
 waymark-worker-remote-pool = { workspace = true }
 waymark-worker-status-reporter = { workspace = true }
