+++ crates/lib/worker-python-bringup/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "waymark-worker-python-bringup"
+edition = "2024"
+version.workspace = true
+publish.workspace = true
+
+[dependencies]
+waymark-worker-process-pool = { workspace = true }
+waymark-worker-python = { workspace = true }
+waymark-worker-remote-bringup = { workspace = true }
+
+thiserror = { workspace = true }
+tokio = { workspace = true, features = ["rt"] }
+tokio-util = { workspace = true }
